+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Assume that conditions are safe if the enclosure engineering mode is enabled.
 * Prevent attempts to close the dome or park the telescopes in local mode.
 * Allow twilight flats with exposure times rounded up to a tenth of a second.
+* Do not fail the twilight flats in the Overwatcher if the exposure time is too long for the last fibre. Instead, just log a warning and mark the calibration as done.
 
 ### 🔧 Fixed